refactor(stress-test): move tests to a separate file
